requirements.txt
sqlite3
selenium
pandas
numpy
requests
python-dotenv
matplotlib